# ms-gateway/docs/swagger.yaml
definitions:
  helper.Message:
    properties:
      message:
        type: string
    type: object
  helper.MessageDetails:
    properties:
      detail:
        type: string
      message:
        type: string
    type: object
...
  model.CreatePaymentRequest:
    properties:
      amount:
        type: integer
      method:
        type: string
    type: object
...
  model.Payment:
    properties:
      amount:
        type: integer
      bookingID:
        type: string
      createdAt:
        type: string
      id:
        type: string
      invoice_id:
        type: string
      invoice_url:
        type: string
      method:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: arviansyahnur3@gmail.com
    name: Arviansyah Nur
    url: http://www.swagger.io/support
  description: An e-commerce API for booking hotel and with many features, made using
    microservices and gRPC.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: vihotel
  version: "1.0"
paths:
...
  /api/v1/payment/:booking_id:
    get:
      consumes:
      - application/json
      description: Create payments
      parameters:
      - description: The input user struct
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CreatePaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Payment
  /api/v1/payment/:payment_id:
    get:
      consumes:
      - application/json
      description: Get payment
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Payment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Message'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.MessageDetails'
      summary: Private
      tags:
      - Payment
...
swagger: "2.0"